requirements.txt
biopython
streamlit
pandas
pytest
